parse empty statements as nil instead of an error
Previously empty statements would result in parse errors. This can
happen in cases where they do not initially appear empty, e.g. if the
contained only comments. For example, something similar to this appears
in `mysqldump` output:

```
/* some comment*/;
/* another comment */;
CREATE TABLE ...
```

`Lex` usually discards comments until it finds a token to return but in
the above case that is nil since it is an empty statement.

This change extends the parser to handle the empty statement (after lex
does its comment discarding in the above example).

Fixes #31.